Refactor LaTeX/PDF branches of tag functions to use X-expressions
Due to the need to process ◊hyperlink tags within other tags, the
ltx/pdf branch of all tag functions need to return tagged
X-expressions, not strings. This is so that when `decode-elements` is
called, the elements will all be valid tagged x-expressions. At the
end, `root` will use `txt-decode` to convert these elements into simple
strings.