Refactor date functions
Changing datestring->date to return a Racket date value (rather than a
string). This a) makes the function's name more accurate, and b) allows
us to use it in generating date strings in other formats, such as
RFC3339 in the Atom feed. A new function `pubdate->english` is used to
supply the functionality previously provided by `datestring->date`