Update markup in Flatland chapters to use inline math
Using the inline math environment gives better results than manually
superscripting numbers to simulate exponents. Note that even when a
single number is being referred to in a mathematics context (e.g. "9"
in connection with "3^3") I place that single number in an inline math
environment also: this ensures the same style of figure is used among
related numbers (old-style figures are used in the body text, but
lining figures in math).