Revise tag functions intended for sectional breaks
Renamed ‘doc-section’ to just ‘section’. Added a ◊pause tag intended to
be used in front of ◊new-thought, because while the ‘\newthought’
command in our LaTeX doc automatically adds extra vertical space before
the containing paragraph, the only way to do this in HTML/CSS involves
JavaScript (CSS has no “element which contains a .newthought”
selector). So the ◊pause tag adds nothing in LaTeX but inserts <hr> in
HTML. Some more advanced Pollen processing would be a better way of
doing this.